+++ eve/tests/renders.py
@@ -338,11 +338,11 @@ class TestRenders(TestBase):
             self.app.config["URL_PREFIX"], self.app.config["API_VERSION"]
         )
 
-        del (self.domain["peopleinvoices"])
-        del (self.domain["peoplerequiredinvoices"])
-        del (self.domain["peoplesearches"])
-        del (self.domain["internal_transactions"])
-        del (self.domain["child_products"])
+        del self.domain["peopleinvoices"]
+        del self.domain["peoplerequiredinvoices"]
+        del self.domain["peoplesearches"]
+        del self.domain["internal_transactions"]
+        del self.domain["child_products"]
         for _, settings in self.app.config["DOMAIN"].items():
             # resource endpoint
             url = "%s/%s/" % (prefix, settings["url"])